# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.HtmlBody:
    properties:
      data:
        additionalProperties: true
        type: object
      html:
        type: string
    type: object
...
  object_builder_service.CreateViewRequest:
    properties:
      app_id:
        type: string
      attributes:
        $ref: '#/definitions/structpb.Struct'
      calendar_from_slug:
        type: string
      calendar_to_slug:
        type: string
      columns:
        items:
          type: string
        type: array
      default_limit:
        type: string
      disable_dates:
        $ref: '#/definitions/object_builder_service.DisableDates'
      group_fields:
        items:
          type: string
        type: array
      is_editable:
        type: boolean
      main_field:
        type: string
      multiple_insert:
        type: boolean
      multiple_insert_field:
        type: string
      name:
        type: string
      project_id:
        type: string
      quick_filters:
        items:
          $ref: '#/definitions/object_builder_service.QuickFilter'
        type: array
      relation_id:
        type: string
      relation_table_slug:
        type: string
      status_field_slug:
        type: string
      table_label:
        type: string
      table_slug:
        type: string
      time_interval:
        type: integer
      type:
        type: string
      updated_fields:
        items:
          type: string
        type: array
      users:
        items:
          type: string
        type: array
      view_fields:
        items:
          type: string
        type: array
    type: object
  object_builder_service.DisableDates:
    properties:
      day_slug:
        type: string
      table_slug:
        type: string
      time_from_slug:
        type: string
      time_to_slug:
        type: string
    type: object
...
  object_builder_service.GetAllViewsResponse:
    properties:
      count:
        type: integer
      views:
        items:
          $ref: '#/definitions/object_builder_service.View'
        type: array
    type: object
...
  object_builder_service.PdfBody:
    properties:
      link:
        type: string
    type: object
  object_builder_service.QuickFilter:
    properties:
      default_value:
        type: string
      field_id:
        type: string
    type: object
...
  object_builder_service.View:
    properties:
      app_id:
        type: string
      attributes:
        $ref: '#/definitions/structpb.Struct'
      calendar_from_slug:
        type: string
      calendar_to_slug:
        type: string
      columns:
        items:
          type: string
        type: array
      default_limit:
        type: string
      disable_dates:
        $ref: '#/definitions/object_builder_service.DisableDates'
      group_fields:
        items:
          type: string
        type: array
      id:
        type: string
      is_editable:
        type: boolean
      main_field:
        type: string
      multiple_insert:
        type: boolean
      multiple_insert_field:
        type: string
      name:
        type: string
      project_id:
        type: string
      quick_filters:
        items:
          $ref: '#/definitions/object_builder_service.QuickFilter'
        type: array
      relation_id:
        type: string
      relation_table_slug:
        type: string
      status_field_slug:
        type: string
      table_label:
        type: string
      table_slug:
        type: string
      time_interval:
        type: integer
      type:
        type: string
      updated_fields:
        items:
          type: string
        type: array
      users:
        items:
          type: string
        type: array
      view_fields:
        items:
          type: string
        type: array
    type: object
...
  structpb.Struct:
    properties:
      fields:
        additionalProperties:
          $ref: '#/definitions/structpb.Value'
        description: Unordered map of dynamically typed values.
        type: object
    type: object
  structpb.Value:
    properties:
      kind:
        description: "The kind of value.\n\nTypes that are assignable to Kind:\n\t*Value_NullValue\n\t*Value_NumberValue\n\t*Value_StringValue\n\t*Value_BoolValue\n\t*Value_StructValue\n\t*Value_ListValue"
    type: object
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /v1/html-to-pdf:
    post:
      consumes:
      - application/json
      description: Convert html to pdf
      operationId: convert_html_to_pdf
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: HtmlBody
        in: body
        name: view
        required: true
        schema:
          $ref: '#/definitions/models.HtmlBody'
      produces:
      - application/json
      responses:
        "201":
          description: PdfBody data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.PdfBody'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Convert html to pdf
      tags:
      - View
...
  /v1/template-to-html:
    post:
      consumes:
      - application/json
      description: Convert template to html
      operationId: convert_template_to_html
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: TemplateBody
        in: body
        name: view
        required: true
        schema:
          $ref: '#/definitions/models.HtmlBody'
      produces:
      - application/json
      responses:
        "201":
          description: HtmlBody data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.HtmlBody'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Convert template to html
      tags:
      - View
...
  /v1/view:
    get:
      consumes:
      - application/json
      description: Get view list
      operationId: get_view_list
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - in: query
        name: app_id
        type: string
      - in: query
        name: project_id
        type: string
      - in: query
        name: table_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ViewBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.GetAllViewsResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get view list
      tags:
      - View
    post:
      consumes:
      - application/json
      description: Create view
      operationId: create_view
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: CreateViewRequestBody
        in: body
        name: view
        required: true
        schema:
          $ref: '#/definitions/object_builder_service.CreateViewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: View data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.View'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create view
      tags:
      - View
    put:
      consumes:
      - application/json
      description: Update view
      operationId: update_view
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: UpdateViewRequestBody
        in: body
        name: view
        required: true
        schema:
          $ref: '#/definitions/object_builder_service.View'
      produces:
      - application/json
      responses:
        "200":
          description: View data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.View'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update view
      tags:
      - View
...
  /v1/view/{view_id}:
    delete:
      consumes:
      - application/json
      description: Delete view
      operationId: delete_view
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: view_id
        in: path
        name: view_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete view
      tags:
      - View
    get:
      consumes:
      - application/json
      description: Get single view
      operationId: get_view_by_id
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: view_id
        in: path
        name: view_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ViewBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.View'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get single view
      tags:
      - View
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"